requirements.txt
spacy==1.8.0
tensorflow=1.1.0
python-igraph
nltk==3.2.1
rdflib
yappi
gensim
more_itertools
parvusdb==0.0.10
hy